DPT_standalone/requirements.txt
numpy==1.20.3
opencv-python==4.5.2.54
Pillow==8.2.0
timm==0.4.9
torch==1.8.1
torchvision==0.9.1
typing-extensions==3.10.0.0
numba==0.53.1
